ai code reviewer/requirements.txt
streamlit
groq
python-dotenv
PyGithub
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
chromadb
sentence-transformers
requests
